ref: Reorganize file structure of cabinet module
