+++ src/modules/cabinet/components/Contacts/CabinetContacts.styled.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { FONTS, MEDIA } from '../../theme';
+import { FONTS, MEDIA } from '../../../theme';
 import expandIcon from '../../../assets/images/icon-expand-more.png';
 
 export const Header = styled.h3`
