+++ src/modules/cabinet/components/Favourites/CabinetFavourites.tsx
