+++ src/modules/cabinet/components/Tabs/CabinetTabs.styled.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { COLORS, FONTS } from '../../theme';
+import { COLORS, FONTS } from '../../../theme';
 
 export const Tabs = styled.ul`
   display: flex;
