+++ src/modules/cabinet/components/Tabs/CabinetTabs.tsx
